refactor(auth): restructure authentication logic and remove unused options
