+++ app/(auth)/sign-in/page.jsx
@@ -53,6 +53,8 @@ export default function SignInForm() {
         password: data.password,
       });
 
+      console.log(result);
+
       if (result?.error) {
         console.error("Sign in error:", result.error);
         toast({
