+++ app/api/auth/[...nextauth]/route.js
@@ -1,6 +1,5 @@
 import NextAuth from "next-auth/next";
-import { authOptions } from "./options";
-
+import { authOptions } from "../../../../lib/authOptions";
 const handler = NextAuth(authOptions);
 
 export { handler as GET, handler as POST };
